Renamed package to be more general
