+++ ihmc-perception/src/main/java/us/ihmc/perception/sceneGraph/rigidBody/doors/DoorNodeTools.java
@@ -2,7 +2,7 @@ package us.ihmc.perception.sceneGraph.rigidBody.doors;
 
 import us.ihmc.euclid.transform.RigidBodyTransform;
 import us.ihmc.perception.detections.PersistentDetection;
-import us.ihmc.perception.detections.YOLOv8.YOLOv8InstantDetection;
+import us.ihmc.perception.detections.yolo.YOLOv8InstantDetection;
 import us.ihmc.perception.sceneGraph.SceneGraph;
 import us.ihmc.perception.sceneGraph.SceneNode;
 import us.ihmc.perception.sceneGraph.modification.SceneGraphModificationQueue;
