+++ ihmc-perception/src/test/java/us/ihmc/perception/detections/yolo/YOLOv8ObjectDetectorTest.java
@@ -1,0 +1,14 @@
+package us.ihmc.perception.detections.yolo;
+
+import org.junit.jupiter.api.Test;
+
+public class YOLOv8ObjectDetectorTest
+{
+   private static final String
+
+   @Test
+   public void testMultipleModels()
+   {
+
+   }
+}
